Move AST fine-tuning command to CLI module
Deleted ast_finetune.py and integrated its functionality as the ast_fintetune command within cli.py. This consolidates command definitions into a single CLI entry point for easier maintenance and usage.
